各种优化

+++ Demon Lord Roguelike/Assets/Scrpits/Bean/Game/GashaponItemBean.cs
@@ -5,7 +5,7 @@ public class GashaponItemBean
 {
     public CreatureBean creatureData;
 
-    public GashaponItemBean(int creatureId)
+    public GashaponItemBean(long creatureId)
     {
         creatureData = new CreatureBean(creatureId);
         creatureData.AddAllSkin();
